+++ src/extension/DragExtension.js
@@ -1,6 +1,14 @@
 (function () {
     "use strict";
 
-    Rise.DragExtension = Rise.Extension.extend();
+    Rise.DragExtension = Rise.Extension.extend({
+        init: function (rise) {
+
+        },
+
+        onElementAdded: function () {
+            console.log('element added');
+        }
+    });
 
 }());